Nuke the Vaccine files
