+++ src/actions/Bluetooth/DownloadActions.js
@@ -1,0 +1,101 @@
+/**
+ * mSupply Mobile
+ * Sustainable Solutions (NZ) Ltd. 2021
+ */
+
+import moment from 'moment';
+import { PermissionActions } from '../PermissionActions';
+import BleService from '../../bluetooth/BleService';
+import TemperatureLogManager from '../../bluetooth/TemperatureLogManager';
+import { VACCINE_CONSTANTS } from '../../utilities/modules/vaccines/index';
+import { UIDatabase } from '../../database/index';
+import { VACCINE_ENTITIES } from '../../utilities/modules/vaccines/constants';
+import { selectIsDownloadingLogs } from '../../selectors/Bluetooth/bluetooth';
+
+export const DOWNLOAD_ACTIONS = {
+  DOWNLOAD_LOGS_START: 'Vaccine/downloadLogsStart',
+  DOWNLOAD_LOGS_ERROR: 'Vaccine/downloadLogsError',
+  DOWNLOAD_LOGS_COMPLETE: 'Vaccine/downloadLogsComplete',
+};
+
+const downloadLogsStart = () => ({
+  type: DOWNLOAD_ACTIONS.DOWNLOAD_LOGS_START,
+});
+const downloadLogsError = () => ({
+  type: DOWNLOAD_ACTIONS.DOWNLOAD_LOGS_ERROR,
+});
+const downloadLogsComplete = () => ({
+  type: DOWNLOAD_ACTIONS.DOWNLOAD_LOGS_COMPLETE,
+});
+
+const downloadAllLogs = () => async dispatch => {
+  dispatch(downloadLogsStart());
+  // Ensure there are some sensors which have been assigned a location before syncing.
+  const sensors = UIDatabase.objects('Sensor').filtered('location != null && isActive == true');
+
+  if (!sensors.length) {
+    // TODO: Should we do something if there are errors?
+    dispatch(downloadLogsError());
+    return null;
+  }
+
+  for (let i = 0; i < sensors.length; i++) {
+    // Intentionally sequential
+    // eslint-disable-next-line no-await-in-loop
+    await dispatch(downloadLogsFromSensor(sensors[i]));
+  }
+  dispatch(downloadLogsComplete());
+  return null;
+};
+
+const downloadLogsFromSensor = sensor => async () => {
+  const { macAddress, logInterval } = sensor;
+
+  const downloadedLogsResult =
+    (await BleService().downloadLogsWithRetries(
+      macAddress,
+      VACCINE_CONSTANTS.MAX_BLUETOOTH_COMMAND_ATTEMPTS
+    )) ?? {};
+
+  if (downloadedLogsResult) {
+    const savedTemperatureLogs = UIDatabase.objects(VACCINE_ENTITIES.TEMPERATURE_LOG)
+      .filtered('sensor.macAddress == $0', macAddress)
+      .sorted('timestamp', true);
+
+    const [mostRecentLog] = savedTemperatureLogs;
+    const mostRecentLogTime = mostRecentLog ? mostRecentLog.timestamp : null;
+    const nextPossibleLogTime = mostRecentLogTime
+      ? moment(mostRecentLogTime).add(logInterval, 's')
+      : 0;
+
+    const numberOfLogsToSave = await TemperatureLogManager().calculateNumberOfLogsToSave(
+      logInterval,
+      nextPossibleLogTime
+    );
+
+    const temperatureLogs = await TemperatureLogManager().createLogs(
+      downloadedLogsResult,
+      sensor,
+      numberOfLogsToSave,
+      mostRecentLogTime
+    );
+
+    await TemperatureLogManager().saveLogs(temperatureLogs);
+  }
+
+  return null;
+};
+
+const startDownloadAllLogs = () => async (dispatch, getState) => {
+  // Ensure there isn't already a download in progress before starting a new one
+  const state = getState();
+  const isDownloadingLogs = selectIsDownloadingLogs(state);
+  if (isDownloadingLogs) return null;
+
+  await PermissionActions.withLocationAndBluetooth(dispatch, getState, downloadAllLogs());
+  return null;
+};
+
+export const DownloadActions = {
+  startDownloadAllLogs,
+};
