+++ src/actions/index.js
@@ -12,6 +12,8 @@ export { WIZARD_ACTIONS, WizardActions } from './WizardActions';
 export { USER_ACTION_TYPES, UserActions } from './UserActions';
 export { CASH_TRANSACTION_ACTION_TYPES, CashTransactionActions } from './CashTransactionActions';
 export { DASHBOARD_ACTION_TYPES, DashboardActions } from './DashboardActions';
-export { VACCINE_ACTIONS, VaccineActions } from './VaccineActions';
+export { DOWNLOAD_ACTIONS, DownloadActions } from './Bluetooth/DownloadActions';
+export { BLINK_ACTIONS, BlinkActions } from './Bluetooth/BlinkActions';
+export { SCAN_ACTIONS, ScanActions } from './Bluetooth/ScanActions';
 
 export * from './Entities';
