--- src/selectors/vaccine.js
@@ -1,9 +1,0 @@
-/**
- * mSupply Mobile
- * Sustainable Solutions (NZ) Ltd. 2021
- */
-
-export const selectIsDownloadingLogs = ({ vaccine }) => {
-  const { isDownloadingLogs = false } = vaccine || {};
-  return isDownloadingLogs;
-};
